Integrate parallel strategies

+++ parallel_jobs_manager.py
@@ -11,6 +11,7 @@ from abc import ABC, abstractmethod
 import subprocess
 import os
 import shutil
+from modules.common import to_log
 from version import __version__
 
 __author__ = "Bogdan M. Kirilenko"
@@ -20,15 +21,18 @@ class ParallelizationStrategy(ABC):
     """
     Abstract base class for a parallelization strategy.
     """
+    def __init__(self):
+        self._process = None
 
     @abstractmethod
-    def execute(self, joblist_path, manager_data, label, **kwargs):
+    def execute(self, joblist_path, manager_data, label, wait=False, **kwargs):
         """
         Execute the jobs in parallel.
 
         :param joblist_path: Path to the joblist file.
         :param manager_data: Data from the manager class.
         :param label: Label for the run.
+        :param wait: Boolean -> controls whether run blocking or not
         """
         pass
 
@@ -40,6 +44,11 @@ class ParallelizationStrategy(ABC):
         :return: Status of the jobs.
         """
         pass
+
+    def terminate_process(self):
+        """Terminates the associated process"""
+        if self._process:
+            self._process.terminate()
 
 
 class NextflowStrategy(ParallelizationStrategy):
@@ -53,6 +62,7 @@ class NextflowStrategy(ParallelizationStrategy):
     CESAR_CONFIG_MEM_TEMPLATE = "${_MEMORY_}"
 
     def __init__(self):
+        super().__init__()
         self._process = None
         self.joblist_path = None
         self.manager_data = None
@@ -65,20 +75,21 @@ class NextflowStrategy(ParallelizationStrategy):
         self.memory_limit = 16
         self.nf_master_script = None
         self.config_path = None
+        self.return_code = None
 
-    def execute(self, joblist_path, manager_data, label, **kwargs):
+    def execute(self, joblist_path, manager_data, label, wait=False, **kwargs):
         """Implementation for Nextflow."""
         # define parameters
         self.joblist_path = joblist_path
         self.manager_data = manager_data
         self.label = label
+        self.memory_limit = int(kwargs.get("memory_limit", 16))
+
         self.nf_project_path = manager_data.get("nextflow_dir", None)  # in fact, contains NF logs
         self.keep_logs = manager_data.get("keep_nf_logs", False)
         self.use_local_executor = manager_data.get("local_executor", False)
-        self.memory_limit = int(kwargs.get("memory_limit", 16))
         self.nf_master_script = manager_data["NF_EXECUTE"]  # NF script that calls everything
         self.nextflow_config_dir = manager_data.get("nextflow_config_dir", None)
-
         self.config_path = self.__create_config_file()
         # create the nextflow process
 
@@ -86,10 +97,18 @@ class NextflowStrategy(ParallelizationStrategy):
         if self.config_path:
             cmd += f" -c {self.config_path}"
 
+        log_dir = manager_data["logs_dir"]
+        os.mkdir(log_dir) if not os.path.isdir(log_dir) else None
         log_file_path = os.path.join(manager_data["logs_dir"], f"{label}.log")
         with open(log_file_path, "w") as log_file:
-            self._process = subprocess.Popen(cmd, shell=True, stdout=log_file, stderr=log_file)
-        pass
+            to_log(f"Parallel manager: pushing job {cmd}")
+            self._process = subprocess.Popen(cmd,
+                                             shell=True,
+                                             stdout=log_file,
+                                             stderr=log_file,
+                                             cwd=self.nf_project_path)
+        if wait:
+            self._process.wait()
 
     def __create_config_file(self):
         """Create config file and return path to it if needed"""
@@ -97,11 +116,13 @@ class NextflowStrategy(ParallelizationStrategy):
             # for local executor, no config file is needed
             return None
         if self.label.startswith(self.CHAIN_JOBS_PREFIX):
-            config_path = os.path.join(self.nextflow_config_dir, self.CHAIN_CONFIG_TEMPLATE_FILENAME)
+            config_path = os.path.abspath(os.path.join(self.nextflow_config_dir,
+                                                       self.CHAIN_CONFIG_TEMPLATE_FILENAME))
             return config_path
         if self.label.startswith(self.CESAR_JOBS_PREFIX):
             # need to craft CESAR joblist first
-            config_template_path = os.path.join(self.nextflow_config_dir, self.CESAR_CONFIG_TEMPLATE_FILENAME)
+            config_template_path = os.path.abspath(os.path.join(self.nextflow_config_dir,
+                                                                self.CESAR_CONFIG_TEMPLATE_FILENAME))
             with open(config_template_path, "r") as f:
                 cesar_config_template = f.read()
             config_string = cesar_config_template.replace(self.CESAR_CONFIG_MEM_TEMPLATE,
@@ -117,19 +138,22 @@ class NextflowStrategy(ParallelizationStrategy):
 
     def check_status(self):
         """Check if nextflow jobs are done."""
+        if self.return_code:
+            return self.return_code
         running = self._process.poll() is None
         if running:
             return None
+        self.return_code = self._process.returncode
         # the process just finished
         # nextflow provides a huge and complex tree of log files
         # remove them if user did not explicitly ask to keep them
-        if not self.keep_logs and self.nf_project_path:
-            # remove nextflow intermediate files
-            shutil.rmtree(self.nf_project_path) if os.path.isdir(self.nf_project_path) else None
+        # if not self.keep_logs and self.nf_project_path:
+        #     # remove nextflow intermediate files
+        #     shutil.rmtree(self.nf_project_path) if os.path.isdir(self.nf_project_path) else None
         if self.config_path and self.label.startswith(self.CESAR_JOBS_PREFIX):
             # for cesar TOGA creates individual config files
             os.remove(self.config_path) if os.path.isfile(self.config_path) else None
-        return self._process.returncode
+        return self.return_code
 
 
 class ParaStrategy(ParallelizationStrategy):
@@ -141,9 +165,11 @@ class ParaStrategy(ParallelizationStrategy):
     """
 
     def __init__(self):
+        super().__init__()
         self._process = None
+        self.return_code = None
 
-    def execute(self, joblist_path, manager_data, label, **kwargs):
+    def execute(self, joblist_path, manager_data, label, wait=False, **kwargs):
         """Implementation for Para."""
         cmd = f"para make {label} {joblist_path} "
         if "queue_name" in kwargs:
@@ -151,21 +177,45 @@ class ParaStrategy(ParallelizationStrategy):
             cmd += f" -q={queue_name} "
         # otherwise use default medium queue
         if "memory_limit" in kwargs:
-            memory_mb = kwargs["memory_mb"]
+            memory_mb = kwargs["memory_limit"] * 1000  # para uses MB instead of GB
             cmd += f" --memoryMb={memory_mb}"
         # otherwise use default para's 10Gb
 
+        log_dir = manager_data["logs_dir"]
+        os.mkdir(log_dir) if not os.path.isdir(log_dir) else None
         log_file_path = os.path.join(manager_data["logs_dir"], f"{label}.log")
         with open(log_file_path, "w") as log_file:
             self._process = subprocess.Popen(cmd, shell=True, stdout=log_file, stderr=log_file)
+        if wait:
+            self._process.wait()
 
     def check_status(self):
         """Check if Para jobs are done."""
+        if self.return_code:
+            return self.return_code
         running = self._process.poll() is None
         if not running:
-            return self._process.returncode
+            self.return_code = self._process.returncode
+            return self.return_code
         else:
             return None
+
+
+class SnakeMakeStrategy(ParallelizationStrategy):
+    """
+    Not implemented class for Snakemake strategy.
+    Might be helpful for users experiencing issues with Nextflow.
+    """
+    def __int__(self):
+        self._process = None
+        self.return_code = None
+        raise NotImplementedError("Snakemake strategy is not yet implemented")
+
+    def execute(self, joblist_path, manager_data, label, wait=False, **kwargs):
+        raise NotImplementedError("Snakemake strategy is not yet implemented")
+
+    def check_status(self):
+        raise NotImplementedError("Snakemake strategy is not yet implemented")
 
 
 class CustomStrategy(ParallelizationStrategy):
@@ -174,12 +224,15 @@ class CustomStrategy(ParallelizationStrategy):
     """
 
     def __init__(self):
+        super().__init__()
         self._process = None
+        self.return_code = None
+        raise NotImplementedError("Custom strategy is not implemented -> pls see documentation")
 
-    def execute(self, joblist_path, manager_data, label, **kwargs):
+    def execute(self, joblist_path, manager_data, label, wait=False, **kwargs):
         """Custom implementation.
 
-        Please provide your implementation of parallel jobs executor.
+        Please provide your implementation of parallel jobs' executor.
         Jobs are stored in the joblist_path, manager_data is a dict
         containing project-wide TOGA parameters.
 
@@ -218,6 +271,7 @@ class ParallelJobsManager:
         :param strategy: The parallelization strategy to use.
         """
         self.strategy = strategy
+        self.return_code = None
 
     def execute_jobs(self, joblist_path, manager_data, label, **kwargs):
         """
@@ -229,10 +283,14 @@ class ParallelJobsManager:
         """
         self.strategy.execute(joblist_path, manager_data, label, **kwargs)
 
-    def check_jobs_status(self):
+    def check_status(self):
         """
         Check the status of the jobs using the specified strategy.
 
         :return: Status of the jobs.
         """
         return self.strategy.check_status()
+
+    def terminate_process(self):
+        """Terminate associated process."""
+        self.strategy.terminate_process()
